refactor: `Post`, `Member`, `Comment` domain package 생성 및 위치 변경
