+++ healthper/src/main/java/umc/healthper/api/PostController.java
@@ -11,8 +11,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
-import umc.healthper.domain.Member;
-import umc.healthper.domain.Post;
+import umc.healthper.domain.member.Member;
+import umc.healthper.domain.post.Post;
 import umc.healthper.dto.post.*;
 import umc.healthper.exception.ExceptionResponse;
 import umc.healthper.service.MemberService;
