+++ healthper/src/main/java/umc/healthper/domain/comment/CommentStatus.java
@@ -1,4 +1,4 @@
-package umc.healthper.domain;
+package umc.healthper.domain.comment;
 
 public enum CommentStatus {
     NORMAL, REMOVED, BLOCKED
