+++ healthper/src/main/java/umc/healthper/domain/comment/CommentType.java
@@ -1,4 +1,4 @@
-package umc.healthper.domain;
+package umc.healthper.domain.comment;
 
 public enum CommentType {
     COMMENT, NESTED
