+++ healthper/src/main/java/umc/healthper/domain/member/Member.java
@@ -1,8 +1,11 @@
-package umc.healthper.domain;
+package umc.healthper.domain.member;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import umc.healthper.domain.post.Post;
+import umc.healthper.domain.RecordJPA;
+import umc.healthper.domain.comment.Comment;
 import umc.healthper.domain.like.PostLike;
 import umc.healthper.global.BaseTimeEntity;
 
@@ -11,7 +14,7 @@ import javax.validation.constraints.NotNull;
 import java.util.*;
 
 import static javax.persistence.CascadeType.*;
-import static umc.healthper.domain.MemberStatus.*;
+import static umc.healthper.domain.member.MemberStatus.*;
 
 @Entity
 @Getter
