+++ healthper/src/main/java/umc/healthper/domain/member/MemberStatus.java
@@ -1,4 +1,4 @@
-package umc.healthper.domain;
+package umc.healthper.domain.member;
 
 public enum MemberStatus {
     NORMAL, RESIGNED, BLOCKED
