+++ healthper/src/main/java/umc/healthper/domain/post/PostStatus.java
@@ -1,4 +1,4 @@
-package umc.healthper.domain;
+package umc.healthper.domain.post;
 
 public enum PostStatus {
     NORMAL, REMOVED, BLOCKED
