+++ healthper/src/main/java/umc/healthper/dto/comment/CommentResponseDto.java
@@ -3,9 +3,9 @@ package umc.healthper.dto.comment;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import umc.healthper.domain.Comment;
-import umc.healthper.domain.CommentStatus;
-import umc.healthper.domain.Member;
+import umc.healthper.domain.comment.Comment;
+import umc.healthper.domain.comment.CommentStatus;
+import umc.healthper.domain.member.Member;
 import umc.healthper.dto.member.MemberInfoDto;
 
 import java.time.LocalDateTime;
