+++ healthper/src/main/java/umc/healthper/dto/comment/NestedCommentResponseDto.java
@@ -1,9 +1,9 @@
 package umc.healthper.dto.comment;
 
 import lombok.Data;
-import umc.healthper.domain.Comment;
-import umc.healthper.domain.CommentStatus;
-import umc.healthper.domain.Member;
+import umc.healthper.domain.comment.Comment;
+import umc.healthper.domain.comment.CommentStatus;
+import umc.healthper.domain.member.Member;
 import umc.healthper.dto.member.MemberInfoDto;
 
 import java.time.LocalDateTime;
