+++ healthper/src/main/java/umc/healthper/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package umc.healthper.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import umc.healthper.domain.Member;
+import umc.healthper.domain.member.Member;
 
 import java.util.Optional;
