+++ healthper/src/main/java/umc/healthper/repository/RecordRepository.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 import umc.healthper.Section;
-import umc.healthper.domain.Member;
+import umc.healthper.domain.member.Member;
 import umc.healthper.domain.RecordJPA;
 import umc.healthper.dto.record.PostRecordReq;
