+++ healthper/src/main/java/umc/healthper/repository/comment/CommentRepository.java
@@ -1,7 +1,7 @@
 package umc.healthper.repository.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import umc.healthper.domain.Comment;
+import umc.healthper.domain.comment.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long>, CommentRepositoryCustom {
 }
