+++ healthper/src/main/java/umc/healthper/repository/comment/CommentRepositoryCustom.java
@@ -1,6 +1,6 @@
 package umc.healthper.repository.comment;
 
-import umc.healthper.domain.Comment;
+import umc.healthper.domain.comment.Comment;
 
 public interface CommentRepositoryCustom {
     void removeComment(Comment comment);
