+++ healthper/src/main/java/umc/healthper/repository/like/PostLikeRepository.java
@@ -1,8 +1,8 @@
 package umc.healthper.repository.like;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import umc.healthper.domain.Member;
-import umc.healthper.domain.Post;
+import umc.healthper.domain.member.Member;
+import umc.healthper.domain.post.Post;
 import umc.healthper.domain.like.PostLike;
 
 import java.util.Optional;
