+++ healthper/src/main/java/umc/healthper/repository/post/PostRepositoryCustom.java
@@ -1,6 +1,6 @@
 package umc.healthper.repository.post;
 
-import umc.healthper.domain.Post;
+import umc.healthper.domain.post.Post;
 
 public interface PostRepositoryCustom {
     void removePost(Post post);
