+++ healthper/src/main/java/umc/healthper/service/CommentService.java
@@ -3,8 +3,8 @@ package umc.healthper.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.healthper.domain.Comment;
-import umc.healthper.domain.CommentStatus;
+import umc.healthper.domain.comment.Comment;
+import umc.healthper.domain.comment.CommentStatus;
 import umc.healthper.exception.comment.CommentAlreadyRemovedException;
 import umc.healthper.exception.comment.CommentNotFoundException;
 import umc.healthper.repository.comment.CommentRepository;
