+++ healthper/src/main/java/umc/healthper/service/PostLikeService.java
@@ -3,8 +3,8 @@ package umc.healthper.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.healthper.domain.Member;
-import umc.healthper.domain.Post;
+import umc.healthper.domain.member.Member;
+import umc.healthper.domain.post.Post;
 import umc.healthper.domain.like.PostLike;
 import umc.healthper.exception.member.MemberNotFoundByIdException;
 import umc.healthper.exception.post.PostNotFoundException;
