+++ healthper/src/main/java/umc/healthper/service/PostService.java
@@ -5,8 +5,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.healthper.domain.Post;
-import umc.healthper.domain.PostStatus;
+import umc.healthper.domain.post.Post;
+import umc.healthper.domain.post.PostStatus;
 import umc.healthper.dto.post.UpdatePostRequestDto;
 import umc.healthper.exception.post.PostAlreadyRemovedException;
 import umc.healthper.exception.post.PostNotFoundException;
