+++ healthper/src/main/java/umc/healthper/service/RecordService.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.healthper.Section;
-import umc.healthper.domain.Member;
+import umc.healthper.domain.member.Member;
 import umc.healthper.domain.RecordJPA;
 import umc.healthper.dto.record.GetCalenderRes;
 import umc.healthper.dto.record.GetRecordRes;
