+++ healthper/src/test/java/umc/healthper/service/MemberServiceTest.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
-import umc.healthper.domain.Member;
+import umc.healthper.domain.member.Member;
 
 import java.util.List;
