+++ healthper/src/test/java/umc/healthper/service/PostServiceTest.java
@@ -6,8 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
-import umc.healthper.domain.Member;
-import umc.healthper.domain.Post;
+import umc.healthper.domain.member.Member;
+import umc.healthper.domain.post.Post;
 import umc.healthper.dto.post.UpdatePostRequestDto;
 
 import static org.assertj.core.api.Assertions.*;
